lace-util/peimage: Don't include flexible length data directory array in OptionalHeader64
If we include this, the bounds check will fail on PE files that have the
optional header size truncated to the actual number of data directories
present.

If we need the data directories later, we should parse them out
seperately, but for now including them in the unstructured nt_data does
no harm.