lace-platform/efi/linux: Bypass the firmware when booting the kernel
The kernels we want to boot are usually not verified using the firmware
db/dbx database but are verfied indirectly by using shim or by being
embedded in a stubble PE image.

As a result they cannot be passed to the firmware's load image when
Secure Boot is enabled. Instead, this commit implement a basic PE
loader for booting the kernel directly.

shim 16(.1) solves this problem by hooking the firmwares image services
and injecting its own verification mechanism, but we need lace-stubble
to work on shim 15.x as well, and as a result we cannot rely on this for
the moment.